refactor(settings): extract General tab into its own component (#332 phase 1 slice 4)
Fourth and largest mechanical slice of the settings monolith
decomposition (after Permissions #387, Vocabulary #389,
Replacements #390). Moves autostart toggle (with stale-LaunchAgent
warning), HUD-overlay toggle, audio-cues toggle, transcription-
threads slider, and first-run-reset button — plus the PTT-hotkey
editor's host-platform glyph plumbing — out of
\`src/routes/settings/+page.svelte\` into a new
\`src/lib/GeneralTab.svelte\`.

What moved:
- ~17 \`\$state\` declarations (autostart x6, hud x3, sound cues x3,
  inference threads x4, first-run reset x2, isMacOS).
- ~10 IPC handlers (load/toggle for each setting, plus the PTT
  platform probe).
- The 5-step \`Promise.all\` parallel mount-time load.
- The autostart plugin imports + the \`platform()\` probe.
- ~190 lines of markup spanning 5 \`<section class="settings-group">\`
  blocks.
- ~190 lines of CSS for the per-tab classes (slider-row,
  settings-row variants, chord, kbd, etc.) — copied into
  GeneralTab's own scoped style block since Svelte's page-
  scoped styles don't reach child-component markup. Once all
  tabs are extracted these can hoist to a shared module; for
  now the duplication is contained per-tab.

Behavioural delta: lazy load — same as the prior three slices.
The 5 General-tab IPCs now fire only when the tab actually
mounts instead of on every Settings open.

Stats:
- \`settings/+page.svelte\` 2182 → 1592 (-590)
- New \`GeneralTab.svelte\` 666 LOC (state + handlers + lifecycle +
  markup + scoped CSS — the CSS is ~190 lines of that)
